Fix version setting

+++ Project4Common.h
@@ -26,7 +26,7 @@
 
 using json = JSON;
 
-const unsigned int VERSION = 1;
+const double VERSION = 1.0;
 const std::string VALID_TYPES[] = {"list", "listResponse", "pull", "pullResponse", "leave"};
 const std::string BASE64_CHARS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 const char BASE64_PAD_CHAR = '=';
